A bit of file restructure and clean up
