+++ strider/openapi.py
